# packages/web-core/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Duck Engine Web Core API",
    "version": "1.0.0",
    "description": "Core web API for Duck Engine (resource-first materials + file assets; scenes/editor next)",
    "contact": {
      "name": "SpaceDucks Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/auth/invite/validate": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Validate an invite token",
        "description": "Public endpoint used by the invitation acceptance page.",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invite is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "invite"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "invite": {
                      "$ref": "#/components/schemas/InvitePublic"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing token"
          },
          "404": {
            "description": "Invalid/expired invite"
          },
          "409": {
            "description": "Already claimed / user already active"
          }
        }
      }
    },
    "/api/auth/invite/claim": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Claim an invite",
        "description": "Public endpoint used by the invitation acceptance page. Sets the password and activates the user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimInviteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invite claimed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload"
          },
          "404": {
            "description": "Invalid/expired invite"
          },
          "409": {
            "description": "Already claimed / user already active"
          }
        }
      }
    },
...
    "/api/admin/users/invites": {
      "get": {
        "tags": [
          "Admin",
          "Users"
        ],
        "summary": "List user invites",
        "description": "Requires SUPER_ADMIN.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of invites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "count"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserInvite"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin",
          "Users"
        ],
        "summary": "Create a user invite",
        "description": "Requires SUPER_ADMIN. The response includes an invite URL that contains a one-time token.\nThere is no email integration; the super admin is expected to share this URL manually.\n",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInviteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created invite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserInviteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "webcore_auth"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "SUPER_ADMIN",
          "ADMIN",
          "USER"
        ]
      },
...
      "UserInvite": {
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "expiresAt",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "claimedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdByUserId": {
            "type": "string"
          },
          "claimedByUserId": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreateUserInviteRequest": {
        "type": "object",
        "required": [
          "email",
          "name"
        ],
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER"
            ],
            "default": "USER"
          },
          "expiresInHours": {
            "type": "integer",
            "description": "Time-to-live for the invite link (default 168 hours)."
          }
        }
      },
      "CreateUserInviteResponse": {
        "type": "object",
        "required": [
          "inviteUrl",
          "expiresAt"
        ],
        "properties": {
          "inviteUrl": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InvitePublic": {
        "type": "object",
        "required": [
          "email",
          "name",
          "role",
          "expiresAt"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ClaimInviteRequest": {
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "Minimum 10 characters."
          },
          "name": {
            "type": "string",
            "description": "Optional display name override."
          }
        }
      },
...
    }
  },
...
}